Narrowed the scope of DirectSegmentAccess to a SegmentContainer (down from StreamSegmentService).
