+++ segmentstore/server/src/main/java/io/pravega/segmentstore/server/SegmentContainer.java
@@ -11,8 +11,9 @@ package io.pravega.segmentstore.server;
 
 import io.pravega.segmentstore.contracts.SegmentProperties;
 import io.pravega.segmentstore.contracts.StreamSegmentStore;
-
+import java.time.Duration;
 import java.util.Collection;
+import java.util.concurrent.CompletableFuture;
 
 /**
  * Defines a Container for StreamSegments.
@@ -25,4 +26,16 @@ public interface SegmentContainer extends StreamSegmentStore, Container {
      * @return A Collection with the SegmentProperties for these segments.
      */
     Collection<SegmentProperties> getActiveSegments();
+
+    /**
+     * Returns a DirectSegmentAccess object that can be used for operating on a particular StreamSegment directly. The
+     * result of this call should only be used for short periods of time (i.e., while processing a single request) and
+     * not be cached for long-term access.
+     *
+     * @param streamSegmentName The name of the StreamSegment to get for.
+     * @param timeout           Timeout for the operation.
+     * @return A CompletableFuture that, when completed normally, will contain the desired result. If the operation
+     * failed, the future will be failed with the causing exception.
+     */
+    CompletableFuture<DirectSegmentAccess> forSegment(String streamSegmentName, Duration timeout);
 }
